% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdc_dc.R
\name{xdc.dc}
\alias{xdc.dc}
\title{Mixture Decomposition}
\usage{
xdc.dc(es, se, k = 2)
}
\arguments{
\item{es, }{the reported effect sizes}

\item{se, }{the reported standard errors}

\item{k, }{the number of mixture components,
default to 2}
}
\value{
output of mixture decomposition, a
data frame with three columns: (mean, sd,
weight), and k rows representing the k
components
}
\description{
Using the histogram structure to approximately
identify the mixture composition that
minimizes the distance with the input
distribution.
}
