% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_histogram.R
\name{xdc.hist.data}
\alias{xdc.hist.data}
\title{Getting the Histogram from Observed Effect
Sizes and Standard Errors}
\usage{
xdc.hist.data(es, se)
}
\arguments{
\item{es, }{a vector of effect sizes}

\item{se, }{a vector of standard errors (es
and se must be equi-length)}
}
\value{
a data frame with two columns:
\itemize{
\item x: same as input boundaries
\item p: probability for each bin
}
}
\description{
The returned probability vector may not sum
to 1. The remaining value represents the
probability falling outside the range
specified in the boundaries.
}
