% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_histogram.R
\name{xdc.hist.mx}
\alias{xdc.hist.mx}
\title{Getting the Histogram from a Mixture}
\usage{
xdc.hist.mx(boundaries, mxs)
}
\arguments{
\item{boundaries, }{an array of (upper) bin
boundaries. This is usually obtained from
the histogram of observed data (xdc.hist.data)}

\item{mxs, }{a data frame with three columns:
(mean, sd, weight), representing the mixture
composition}
}
\value{
a data frame with two columns:
\itemize{
\item x: same as input boundaries
\item p: probability for each bin
}
}
\description{
This function supports an arbitrary number of
components. The returned probability vector
may not sum to 1. The remaining value
represents the probability falling outside
the range specified in the boundaries
}
