% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdc_ma.R
\name{xdc.ma}
\alias{xdc.ma}
\title{Mixture-based Moderator Estimation}
\usage{
xdc.ma(es, se, ml, k = 2, method = "RMA")
}
\arguments{
\item{es, }{the reported effect sizes}

\item{se, }{the reported standard errors}

\item{ml, }{the independent variables}

\item{k, }{the number of mixture components,
default to 2}

\item{method, }{the method for moderator test,
default to meta-regression, possible
values include:
\itemize{
\item RMA: meta-regression,
\item GRP: subgroup comparison, using
abs(mean1-mean2)/sqrt(se1^2+se2^2)
}}
}
\value{
output of moderator analysis
}
\description{
Using the latent mixture based method to
estimate a moderating effect.
}
