% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdc.R
\docType{package}
\name{xdc}
\alias{xdc}
\title{xdc}
\description{
A package for decomposing a Gaussian mixture distribution with potentially overlapping components
}
\section{Key functions}{

The xdc package provides four categories of functions:
\itemize{
\item xdc.ma, using the latent mixture-based method for moderator estimation
\item xdc.dc, decomposing a Gaussian mixture
\item xdc.plot.mx, plotting one or more mixture distributions (and comparing it with other data distributions)
\item xdc.hist.mx and xdc.hist.data, computing the histogram of a mixture distribution or from the observed data.
Note that histogram is the focal structure we use to decompose a mixture distribution
}
}

